Splitting all inner componnents
